stori_backend/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-rest-knox>=4.2.0
django-cors-headers>=4.3.0
python-decouple>=3.8
psycopg2-binary>=2.9.9
celery>=5.3.0
redis>=5.0.0

# Analysis dependencies
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0
xlrd>=2.0.1
pdfplumber>=0.10.0
PyPDF2>=3.0.0

# OCR and Face Matching
deepface>=0.0.79
tensorflow>=2.13.0
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
Pillow>=10.0.0
mtcnn>=0.1.1
retina-face>=0.0.13

# Machine Learning (Credit Scoring)
lightgbm>=4.1.0
joblib>=1.3.0
scikit-learn>=1.3.0

# Utilities
requests>=2.31.0
python-dateutil>=2.8.2
pytz>=2023.3

# Note: concurrent.futures is built-in Python module (no installation needed)

